pyproject.toml
[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

target-version = "py38"

include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"

docstring-code-format = false

docstring-code-line-length = "dynamic"

[tool.poetry]
name = "gitbenchmark"
version = "0.0.0"
description = "Automatisation du versioning, changelog et release sur GitHub"
authors = ["KalvinVilla <villa.kalvin@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
gitbenchmark = "main:main"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "1.0.1"
requests = "2.32.3"
toml = "0.10.2"
rich = "13.7.1"
pydantic = "2.9.1"
colorlog = "6.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
